requirements.txt
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3